add validation form

+++ src/ts/service/vehicleService.ts
@@ -8,60 +8,65 @@ import { createVehicleID } from "../utils/vehicleUtils";
 import { addVehicleToRentModal, removeVehicleFromRentModal } from "../utils/vehicleUtils";
 import type { Category } from "../models/Category";
 import type { Car } from "../models/Car";
+import { validateVehicleForm } from "../validation/vehicleValidation";
 
 export function initVehicleCreation(): void {
-    const addVehicleForm = document.getElementById("addVehicleForm") as HTMLFormElement;
-    if (!addVehicleForm) {
-        console.error("Форма не знайдена");
-        return;
-    }
-    addVehicleForm.addEventListener("submit", async (event) => {
-        event.preventDefault();
+  const addVehicleForm = document.getElementById("addVehicleForm") as HTMLFormElement;
+  if (!addVehicleForm) {
+    console.error("Форма не знайдена");
+    return;
+  }
+  addVehicleForm.addEventListener("submit", async (event) => {
+    event.preventDefault();
 
-        const formData = new FormData(addVehicleForm);
-        const type: Category = formData.get("vechicle-type") as Category;
-        const plateNumber: string = formData.get("plate-number")?.toString().trim() || "";
-        const dateOfStartRental: Date = new Date(formData.get("date-of-start-rental") as string);
-        const rentalDuration: number = parseInt(formData.get("rental-duration") as string, 10);
-        const price: number = parseFloat(formData.get("price-rental") as string);
-    
-        const selectedCarId = (formData.get("car") as string)?.trim();
-        if (!selectedCarId) {
-            console.warn("ID автомобіля не вибрано.");
-            return;
-        }
-    
-        let carDb: Car | null = null;
-        try {
-            carDb = await FirebaseService.getCarById(selectedCarId) as Car;
-            carDb.setId = selectedCarId; 
-            if (!carDb) {
-                console.warn("Автомобіль не знайдено за ID:", selectedCarId);
-                return;
-            }
-        } catch (error) {
-            console.error("Помилка при завантаженні автомобіля:", error);
-            return;
-        }
-    
-        if (!type || !carDb || !plateNumber || isNaN(rentalDuration) || !dateOfStartRental || isNaN(price)) {
-            console.warn("Будь ласка, заповніть усі поля коректно.");
-            return;
-        }
-    
-        const newVehicleDto: VehicleDTO = new VehicleDTO(carDb, type, plateNumber, dateOfStartRental, rentalDuration, price);
-        const newVehicle: Vehicle = new Vehicle(newVehicleDto);
-    
-        try {
-            await FirebaseService.addVehicle(newVehicle);
-            console.log("Транспортний засіб успішно додано:", newVehicle);
-            renderVehicleBlock(newVehicle);
-            addVehicleForm.reset();
-            
-        } catch (error) {
-            console.error("Помилка при додаванні транспортного засобу:", error);
-        }
-    });
+    const formData = new FormData(addVehicleForm);
+    const type: Category = formData.get("vechicle-type") as Category;
+    const plateNumber: string = formData.get("plate-number")?.toString().trim() || "";
+    const dateOfStartRental: Date = new Date(formData.get("date-of-start-rental") as string);
+    const rentalDuration: number = parseInt(formData.get("rental-duration") as string, 10);
+    const price: number = parseFloat(formData.get("price-rental") as string);
+
+    if (!validateVehicleForm(addVehicleForm)) {
+      return;
+    }
+
+    const selectedCarId = (formData.get("car") as string)?.trim();
+    if (!selectedCarId) {
+      console.warn("ID автомобіля не вибрано.");
+      return;
+    }
+
+    let carDb: Car | null = null;
+    try {
+      carDb = await FirebaseService.getCarById(selectedCarId) as Car;
+      carDb.setId = selectedCarId;
+      if (!carDb) {
+        console.warn("Автомобіль не знайдено за ID:", selectedCarId);
+        return;
+      }
+    } catch (error) {
+      console.error("Помилка при завантаженні автомобіля:", error);
+      return;
+    }
+
+    if (!type || !carDb || !plateNumber || isNaN(rentalDuration) || !dateOfStartRental || isNaN(price)) {
+      console.warn("Будь ласка, заповніть усі поля коректно.");
+      return;
+    }
+
+    const newVehicleDto: VehicleDTO = new VehicleDTO(carDb, type, plateNumber, dateOfStartRental, rentalDuration, price);
+    const newVehicle: Vehicle = new Vehicle(newVehicleDto);
+
+    try {
+      await FirebaseService.addVehicle(newVehicle);
+      console.log("Транспортний засіб успішно додано:", newVehicle);
+      renderVehicleBlock(newVehicle);
+      addVehicleForm.reset();
+
+    } catch (error) {
+      console.error("Помилка при додаванні транспортного засобу:", error);
+    }
+  });
 }
 
 export function renderVehicleBlock(vechicle: Vehicle): void {
@@ -106,31 +111,36 @@ export function renderVehicleBlock(vechicle: Vehicle): void {
     dialog.openDialog();
     const form = document.getElementById("editVehicleForm") as HTMLFormElement;
     if (!form) return;
-  
-  
+
+
     (document.getElementById("editPlateNumber") as HTMLInputElement).value = vechicle.getPlateNumber;
     (document.getElementById("editStartDate") as HTMLInputElement).value = vechicle.getStartDate.toISOString().split('T')[0];
     (document.getElementById("editRentalDuration") as HTMLInputElement).value = vechicle.getRentalDuration.toString();
     (document.getElementById("editPriceVehicle") as HTMLInputElement).value = vechicle.getPrice.toString();
-    console.log(vechicle.getPrice.toString())
+
+
+
     form.onsubmit = async (event) => {
       event.preventDefault();
-  
+
+      if (!validateVehicleForm(form)) {
+        return;
+      }
       const vehicleDTO: VehicleDTO | null = await readVehicleFormData(form);
       if (vehicleDTO !== null) {
         const updatedVehicle = new Vehicle(vehicleDTO);
         updatedVehicle.setId = vechicle.getId;
-  
+
         const newId = createVehicleID(vechicle.getPlateNumber, vechicle.getStartDate, vechicle.getRentalDuration);
         try {
           await FirebaseService.deleteVehicle(vechicle.getId);
           updatedVehicle.setId = newId;
-  
+
           await FirebaseService.setVehicle(updatedVehicle, vechicle.getId, newId);
           block.remove();
           removeVehicleFromRentModal(vechicle.getId);
           renderVehicleBlock(updatedVehicle);
-  
+
           dialog.closeDialog();
         } catch (error) {
           console.error("Помилка при оновленні транспортного засобу:", error);
